fixed the chess promotion, changed the uint32_t to uint16_t

+++ src/cpp/include/chess_game.hpp
@@ -10,8 +10,8 @@
 #include "chess.hpp" // Disservin single-header library
 
 namespace az73 {
-[[nodiscard]] std::uint32_t encode(const chess::Board& board, const chess::Move& mv);
-[[nodiscard]] chess::Move decode_action(std::uint32_t action, const chess::Board& board);
+[[nodiscard]] std::uint16_t encode(const chess::Board& board, const chess::Move& mv);
+[[nodiscard]] chess::Move decode_action(std::uint16_t action, const chess::Board& board);
 } // namespace az73
 
 class ChessGame {
@@ -22,9 +22,9 @@ public:
 
     [[nodiscard]] const chess::Board& currentBoard() const noexcept { return history_.back(); }
     void makeMove(const chess::Move& m);
-    void makeMove(const std::uint32_t m);
+    void makeMove(const std::uint16_t m);
     [[nodiscard]] Tensor encodeTensor() const;
-    [[nodiscard]] std::vector<std::uint32_t> legalMoves() const;
+    [[nodiscard]] std::vector<std::uint16_t> legalMoves() const;
 
 private:
     std::vector<chess::Board> history_;
